pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbtopo"
version = "0.1.0"
description = "Load IGN BD TOPO into Databricks Delta tables"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "tqdm>=4.66",
    "py7zr>=0.21",
    "geopandas>=0.14",
    "fiona>=1.9",
    "shapely>=2.0",
    "pyproj>=3.6",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.scripts]
dbtopo = "dbtopo.cli:main"

[project.entry-points."databricks"]
download = "dbtopo.cli:download"
load = "dbtopo.cli:load"
validate = "dbtopo.cli:validate"

[tool.hatch.build.targets.wheel]
packages = ["src/dbtopo"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
